itera_lib/requirements.txt
click>=8.0.0
rich>=10.0.0
GitPython>=3.1.0
openai>=1.0.0
pydantic>=2.0.0
faiss-cpu>=1.7.0
networkx>=2.6.0
matplotlib>=3.4.0 